requirements.txt
absl-py==1.4.0
astunparse==1.6.3
batchgenerators==0.25
blessed==1.20.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.0.7
cycler==0.11.0
einops==0.6.1
et-xmlfile==1.1.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.39.4
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==1.1
grpcio==1.54.2
h5py==3.9.0
huggingface-hub==0.15.1
idna==3.4
imageio==2.28.1
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
linecache2==1.0.0
lit==15.0.7
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
natsort==8.3.1
networkx==3.1
nibabel==5.1.0
numpy==1.24.3
nvidia-ml-py==11.525.112
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
openTSNE==1.0.0
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
traceback2==1.4.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
unittest2==1.1.0
urllib3==1.26.13
vit-pytorch==1.2.2
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.15.0
